==> notebooks/04_identifying_collocations.ipynb <==
# Identifying collocations

<br>

**Language: Python**

This notebook looks at frequency statistics from COCA to identify potential collocations to change in the original and normalized text versions. See dissertation section 6.1.1.1.

**Notebook contents:**
1. [Initial setup](#1.-Initial-setup)
2. [Find collocations](#2.-Find-collocations)

## 1. Initial setup

# Import necessary modules

import pandas as pd
import pprint
from IPython.core.interactiveshell import InteractiveShell
import joblib

# Set preferred notebook format

%pprint # Turn off pretty printing
InteractiveShell.ast_node_interactivity = "all" # Show all output, not just last item
pd.set_option('display.max_columns', 999) # Allow viewing of all columns

**Note:** As described in the [README.md]('../README.md'), The frequency information from COCA referenced here is not freely available but can be purchased at https://corpus.byu.edu/coca. Without this data you will be able to see a few rows of these dataframes, but will not be able to run the code yourself.

# Load COCA collocation dataframe

col_df = joblib.load('../../COCA_data/COCA_2020_collocation_df.pkl')
col_df.head()

# Load texts for analysis (Original and Normalized)

B1_df = joblib.load('../docs/B1_cols.pkl')
B2_df = joblib.load('../docs/B2_cols.pkl')
C1_df = joblib.load('../docs/C1_cols.pkl')

# Combine the three dfs

texts = pd.concat([B1_df, B2_df, C1_df]).reset_index(drop=True)
texts

## 2. Find collocations

# Create set of all collocations in COCA

all_cols = set(col_df.collocation.to_list())
len(all_cols)

# Create column of cols in each text

texts['cols'] = texts.possible_cols.apply(lambda row: [x for x in row if x in all_cols])
texts

### Prompt collocations

# Prepare prompt text

prompt = "Children who are brought up in families that do not have large amounts of money are better prepared to deal with the problems of adult life than children brought up by wealthy parents. To what extent do you agree or disagree with this opinion?"
prompt = prompt.lower().replace('.','').replace('?','')
prompt

# Tag with CLAWS7 at http://ucrel-api.lancaster.ac.uk/claws/free.html

prompt_POS = 'children_NN2 who_PNQS are_VBR brought_VVN up_RP in_II families_NN2 that_CST do_VD0 not_XX have_VHI large_JJ amounts_NN2 of_IO money_NN1 are_VBR better_RRR prepared_VVN to_TO deal_VVI with_IW the_AT problems_NN2 of_IO adult_NN1 life_NN1 than_CSN children_NN2 brought_VVN up_RP by_II wealthy_JJ parents_NN2 to_II what_DDQ extent_NN1 do_VD0 you_PPY agree_VVI or_CC disagree_VVI with_IW this_DD1 opinion_NN1'
prompt_POS = [x.split('_') for x in prompt_POS.split(' ')]
prompt_POS = [(x[0],x[1][0].lower()) for x in prompt_POS]
prompt_POS

# Manually lemmatize due to short length

prompt_POS_lem = [('child', 'n'), ('who', 'p'), ('be', 'v'), ('bring', 'v'), ('up', 'r'), ('in', 'i'), ('family', 'n'), ('that', 'c'), ('do', 'v'), ('not', 'x'), ('have', 'v'), ('large', 'j'), ('amount', 'n'), ('of', 'i'), ('money', 'n'), ('be', 'v'), ('better', 'r'), ('prepare', 'v'), ('to', 't'), ('deal', 'v'), ('with', 'i'), ('the', 'a'), ('problem', 'n'), ('of', 'i'), ('adult', 'n'), ('life', 'n'), ('than', 'c'), ('child', 'n'), ('bring', 'v'), ('up', 'r'), ('by', 'i'), ('wealthy', 'j'), ('parent', 'n'), ('to', 'i'), ('what', 'd'), ('extent', 'n'), ('do', 'v'), ('you', 'p'), ('agree', 'v'), ('or', 'c'), ('disagree', 'v'), ('with', 'i'), ('this', 'd'), ('opinion', 'n')]

# Extract potential collocations in span 4

def find_cols(lemma_list):
    col_list = list(zip(lemma_list,lemma_list[1:]))+list(zip(lemma_list,lemma_list[2:]))\
    +list(zip(lemma_list,lemma_list[3:]))+list(zip(lemma_list,lemma_list[4:]))
    return col_list

prompt_cols = find_cols(prompt_POS_lem)

# Remove prompts cols from dataframe cols

texts.possible_cols = texts.possible_cols.apply(lambda row: [x for x in row if x not in prompt_cols])

### Investigate the automated extracted collocations

%pprint

# B1 original

len(set(texts.iloc[0,3]))
set(texts.iloc[0,3])

# And if we consider only ones made up of n, v, adj, adv, or preps

lexical = ['n','v','r','j','i']

set([x for x in texts.iloc[0,3] if x[0][1] in lexical and x[1][1] in lexical])

# B1 normalized

len(set(texts.iloc[1,3]))
set(texts.iloc[1,3])

# B2 original

len(set(texts.iloc[2,3]))
set([x for x in texts.iloc[2,3] if x[0][1] in lexical and x[1][1] in lexical])

# C1 original and normalized (the same collocations)

len(set(texts.iloc[4,3]))
set([x for x in texts.iloc[4,3] if x[0][1] in lexical and x[1][1] in lexical])

[Back to top](#Identifying-collocations)
