==> notebooks/11_processing_qualitative_survey_data.ipynb <==
# Processing qualitative survey data

<br>

**Language: Python**

This notebook shows the process used for cleaning the qualitative data from the Qualtrics survey (rationales for ratings) for future analysis. Prior to being analyzed in this notebook, the comments were manually annotated with tags (see dissertation section 6.2.4 for this details of this scheme).

**Notebook contents:**
- [Initial setup](#Initial-setup)
- [Data processing](#Data-processing)
- [Data analysis](#Data-analysis)

## Initial setup

# Import necessary modules

import pandas as pd
import pprint
from IPython.core.interactiveshell import InteractiveShell
import joblib
import csv
import collections

# Set preferred notebook format

InteractiveShell.ast_node_interactivity = "all" # Show all output, not just last item
pd.set_option('display.max_columns', 999) # Allow viewing of all columns

# Read in survey with annotations

comments = pd.read_csv('../docs/survey_annotated_comments.csv')
comments.head()

## Data processing

# Remove NaN rows and reset index

len(comments)
comments = comments.loc[comments.tags.isna() == False]
comments = comments.reset_index(drop=True)
len(comments)

# Split tags

comments.tags = comments.tags.apply(lambda x: x.split(";"))
comments.head()

# Check and standardize tags

sorted(list(set([x for y in comments.tags for x in y])))

# Create dict

tags_dict = {
    ' FS':'FS',
    ' FS_accuracy':'FS',
    ' accuracy':'accuracy',
    ' accuracy(collocation)':'accuracy(collocation)',
    ' accuracy(word_choice)':'accuracy',
    ' appropriacy':'appropriacy',
    ' coherence':'coherence',
    ' collocation':'collocation',
    ' collocation_accuracy':'accuracy(collocation)',
    ' control':'appropriacy',
    ' diversity':'range',
    ' error_severity':'error_gravity',
    ' fluency':'fluency',
    ' linking_words':'linking_words',
    ' message':'appropriacy',
    ' preposition':'preposition',
    ' register':'style',
    ' sophistication':'sophistication',
    ' spelling':'spelling',
    ' style':'style',
    ' task_adequacy':'appropriacy',
    ' tone':'style',
    ' word_choice':'accuracy(word_choice)',
    ' word_formation':'word_formation',
    'collocation_accuracy':'accuracy(collocation)',
    'task_adequacy':'appropriacy',
    'diversity':'range',
    'register':'style',
}

# Apply dict

comments.tags = comments.tags.apply(lambda row: [tags_dict[x] if x in tags_dict else x for x in row])
sorted(list(set([x for y in comments.tags for x in y])))

## Data analysis

# Counts of each tag

tags = [x for y in comments.tags for x in y]
tag_counts = collections.Counter(tags)
tag_counts

tag_counts.items()

These counts will be used for the visualization created in R in notebook 15.

[Back to top](#Processing-qualitative-survey-data)
