==> notebooks/12_integrating_data_from_FACETS.ipynb <==
# Integrating data from FACETS

<br>

**Language: Python**

This notebook shows the process for reading in the output from the FACETS program, i.e., the MFRM models, and integrating it with the existing survey data. See dissertation sections 4.3 and 6.2.1 for description of MFRM models and these data in particular. The final dataframe is then split into the control data and the main experimental data. 

**Notebook contents:**
- [Initial setup](#Initial-setup)
- [MFRM integration](#MFRM-integration)
- [Split into main and control dataframes](#Split-into-main-and-control-dataframes)
- [Check main ratings data](#Check-main-ratings-data)

## Initial setup

# Import necessary modules

import pandas as pd
import pprint
from IPython.core.interactiveshell import InteractiveShell
import csv
import joblib
import math
import numpy as np

# Set preferred notebook format

%pprint # Turn off pretty printing
InteractiveShell.ast_node_interactivity = "all" # Show all output, not just last item
pd.set_option('display.max_columns', 999) # Allow viewing of all columns
pd.options.mode.chained_assignment = None # Suppress SettingWithCopyWarning

# Read in ratings_long data

ratings_long = joblib.load('../docs/ratings_long.pkl')
ratings_long.head(3)

## MFRM integration

# Read in mfrm data (from Facets)

mfrm = pd.read_csv("../docs/mfrm_output.csv")
mfrm.head()

# Keep only relevant columns

mfrm = mfrm[['Exp','text_id','rater_id','band']]
mfrm = mfrm.rename(columns={'Exp':'fair_score'})

# Convert score to appropriate scale

mfrm.fair_score = (mfrm.fair_score/10).round(1)
mfrm.head()

# Change format so all scores in one cell

mfrm['band_score'] = list(zip(mfrm.band,mfrm.fair_score))

mfrm = (mfrm.groupby(['text_id','rater_id'])
      .agg({'band_score': lambda x: x.tolist()})
      .reset_index())

mfrm.head(10)

# Create temporary columns of text_id and rater_id for easier mapping

mfrm['temp'] = list(zip(mfrm.text_id,mfrm.rater_id))
ratings_long['temp'] = list(zip(ratings_long.text_id,ratings_long.rater_id))

# Items with missing data (removed as 'unexpected responses' in Facets)

mfrm.loc[(mfrm.band_score.str.len() < 5),:].head()
len(mfrm.loc[(mfrm.band_score.str.len() < 5),:])

# Add NA values for any missing data (this is the code for missing data in R)

# Create function
def add_missing(score_list):
    score_list = score_list[:5] + ['x' for x in range(5 - len(score_list))] # Adding padding
    bands = ['TR','CC','LR','GRA','HOL']
    i = 0
    for i in range(len(score_list)):
        if score_list[i][0] != bands[i]:
            score_list.insert(i,(bands[i],np.nan))
            i =+ 1
    return score_list[:5]

# Apply function to band score column and check one example

mfrm.band_score = mfrm.band_score.apply(add_missing)
list(mfrm.loc[(mfrm.text_id == 'Text28') & (mfrm.rater_id == 'R45')].band_score)

# Create dictionary for mapping

mfrm_dict = pd.Series(mfrm.band_score.values,mfrm.temp).to_dict()

# Map to ratings_long df

ratings_long['fair'] = ratings_long.temp.map(mfrm_dict)
del ratings_long['temp']

# Keep only raters used in the MFRM model

ratings_long = ratings_long.loc[~ratings_long.fair.isnull()]

# Create new 'fair' columns

ratings_long['TR_fair'] = [x[0][1] for x in ratings_long['fair']]
ratings_long['CC_fair'] = [x[1][1] for x in ratings_long['fair']]
ratings_long['LR_fair'] = [x[2][1] for x in ratings_long['fair']]
ratings_long['GRA_fair'] = [x[3][1] for x in ratings_long['fair']]
ratings_long['HOL_fair'] = [x[4][1] for x in ratings_long['fair']]

del ratings_long['fair']

ratings_long.loc[ratings_long.text_id == 'Text31']

# Impute NA scores - fair scores are the same for all identical text_id regardless of the rater

# Create df with full scores (no NAs), one row each text_id
ratings_long = ratings_long.sort_values('text_id').reset_index(drop=True)
full = ratings_long[['text_id','TR_fair','CC_fair','LR_fair','GRA_fair','HOL_fair']].dropna().drop_duplicates('text_id')

# Create dictionaries of the full scores
full_TR_dict = pd.Series(full.TR_fair.values,full.text_id).to_dict()
full_CC_dict = pd.Series(full.CC_fair.values,full.text_id).to_dict()
full_LR_dict = pd.Series(full.LR_fair.values,full.text_id).to_dict()
full_GRA_dict = pd.Series(full.GRA_fair.values,full.text_id).to_dict()
full_HOL_dict = pd.Series(full.HOL_fair.values,full.text_id).to_dict()

# Apply dicts to replace NaN values
ratings_long.TR_fair = ratings_long.text_id.map(full_TR_dict)
ratings_long.CC_fair = ratings_long.text_id.map(full_CC_dict)
ratings_long.LR_fair = ratings_long.text_id.map(full_LR_dict)
ratings_long.GRA_fair = ratings_long.text_id.map(full_GRA_dict)
ratings_long.HOL_fair = ratings_long.text_id.map(full_HOL_dict)

## Split into main and control dataframes

Create mini dataframe comprised of the original, normalized, and base texts.

# Separate out base texts

base_texts = ratings_long.loc[ratings_long.text_id.isin(['Text1','Text11','Text21'])]
base_texts['type'] = 'base'
base_texts = base_texts[['type','text_id','rater_id','CEFR','TR','CC','LR','GRA','TR_fair','CC_fair','LR_fair','GRA_fair']]

# Round all ratings down to integer (to match original texts rating options)

base_texts[['TR','CC','LR','GRA','TR_fair','CC_fair','LR_fair','GRA_fair']] = \
base_texts[['TR','CC','LR','GRA','TR_fair','CC_fair','LR_fair','GRA_fair']].apply(np.floor)

# Create 'overall' score and 'overall_fair' score columns

def round_down(x, a):
    return math.floor(x / a) * a

base_texts['overall'] = base_texts[['TR','CC','LR','GRA']].mean(axis=1)
base_texts.overall = base_texts.overall.apply(lambda x: round_down(x,0.5))

base_texts['overall_fair'] = base_texts[['TR_fair','CC_fair','LR_fair','GRA_fair']].mean(axis=1)
base_texts.overall_fair = base_texts.overall_fair.apply(lambda x: round_down(x,0.5))

# Separate out normalized texts

norm_texts = ratings_long.loc[ratings_long.text_id.isin(['Text31','Text32','Text33'])]
norm_texts['type'] = 'norm'
norm_texts = norm_texts[['type','text_id','rater_id','CEFR','TR','CC','LR','GRA','TR_fair','CC_fair','LR_fair','GRA_fair']]

norm_texts

# Round all ratings down to integer (to match original texts rating options)

norm_texts[['TR','CC','LR','GRA','TR_fair','CC_fair','LR_fair','GRA_fair']] = \
norm_texts[['TR','CC','LR','GRA','TR_fair','CC_fair','LR_fair','GRA_fair']].apply(np.floor)

# Create 'overall' score and 'overall_fair' score columns

norm_texts['overall'] = norm_texts[['TR','CC','LR','GRA']].mean(axis=1)
norm_texts.overall = norm_texts.overall.apply(lambda x: round_down(x,0.5))

norm_texts['overall_fair'] = norm_texts[['TR_fair','CC_fair','LR_fair','GRA_fair']].mean(axis=1)
norm_texts.overall_fair = norm_texts.overall_fair.apply(lambda x: round_down(x,0.5))

# Create original texts series

orig_B1 = pd.Series(['orig','IELTS_B1','IELTS','B1',4,4,4,4,4,4,4,4,4,4])
orig_B2 = pd.Series(['orig','IELTS_B2','IELTS','B2',7,6,7,6,7,6,7,6,6.5,6.5])
orig_C1 = pd.Series(['orig','IELTS_C1','IELTS','C1',8,8,8,8,8,8,8,8,8,8])

orig_texts = pd.DataFrame([orig_B1,orig_B2,orig_C1])
orig_texts.columns = norm_texts.columns
orig_texts

# Combine into one dataframe

control = pd.concat([base_texts,norm_texts,orig_texts]).sort_values('CEFR').reset_index(drop=True)
control.head()

control.groupby('text_id').mean()

### Add other text indices from the normalization notebook

# Load original and normalized texts and map to control dataframe

B1_orig_and_norm = joblib.load('../docs/B1_orig&norm.pkl')
B1_orig_and_norm.text_id = pd.Series(['IELTS_B1','Text31'])

B2_orig_and_norm = joblib.load('../docs/B2_orig&norm.pkl')
B2_orig_and_norm.text_id = pd.Series(['IELTS_B2','Text32'])

C1_orig_and_norm = joblib.load('../docs/C1_orig&norm.pkl')
C1_orig_and_norm.text_id = pd.Series(['IELTS_C1','Text33'])

orig_and_norm_texts = pd.concat([B1_orig_and_norm,B2_orig_and_norm,C1_orig_and_norm])
orig_and_norm_texts = orig_and_norm_texts[['text_id','text_len','MLC','CNC','vocD','AG',
                                           'bigram_range','mean_MI', 'mean_tscore','absent_prop','col_errors_per_100',
                                           'correct_cols_per_100','K1to2_p','K3to9_p','K10to16_p']]
orig_and_norm_texts

# Load base texts and map to control dataframe

base_texts = joblib.load('../docs/base_texts_processed.pkl')
base_texts.text_id = pd.Series(['Text1','Text11','Text21'])

base_texts['correct_cols_per_100'] = (base_texts['correct_cols']/base_texts['text_len'])*100
base_texts['col_errors_per_100'] = (base_texts['col_errors']/base_texts['text_len'])*100

base_texts = base_texts[['text_id','text_len','MLC','CNC','vocD','AG','bigram_range',
                         'mean_MI','mean_tscore','absent_prop','col_errors_per_100','correct_cols_per_100',
                         'K1to2_p', 'K3to9_p','K10to16_p']]
base_texts

# Join all texts

all_texts = pd.concat([base_texts,orig_and_norm_texts])
all_texts

# Combine with ratings

control = control.merge(all_texts.set_index('text_id'), on = 'text_id', how = 'outer')
control.head()
len(control)

# Remove control data from ratings_long

len(ratings_long)
ratings_long = ratings_long.loc[~ratings_long.text_id.isin(['Text31','Text32','Text33'])]
len(ratings_long)

# Replace all nan with NA (code for missing data in R)

ratings_long = ratings_long.replace(np.nan,'NA')
control = control.replace(np.nan,'NA')

# Write out control and main ratings to csv for analysis in R

ratings_long.to_csv('../docs/ratings_R.csv',index=False)
control.to_csv('../docs/control_R.csv',index=False)

[Back to top](#Integrating-data-from-FACETS)
